Toast re-arrangements
